pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "online-farm-market"
version = "0.1.0"
description = "A CLI application for managing an online farm market"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "sqlalchemy>=1.4.0",
    "alembic>=1.7.0",
    "python-dotenv>=0.19.0",
    "typer>=0.4.0",
    "rich>=10.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "pytest-mock>=3.6.0",
    "black>=21.0",
    "isort>=5.0.0",
    "flake8>=3.9.0"
]

[project.scripts]
farm-market = "online_farm_market.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
max-complexity = 10
select = "B,C,E,F,W,T4,B9"
